+++ app/Http/Controllers/API/EventReviewAPIController.php
@@ -6,17 +6,18 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 
 use Illuminate\Http\Request;
-use App\Http\Services\EventService;
+use App\Http\Services\EventReviewService;
 
 use App\Photo;
 use App\Event;
 
 class EventReviewAPIController extends Controller
 {
-  private $eventService;
+  private $eventReviewService;
+
   public function __construct()
   {
-      $this->eventService = new EventService();
+      $this->eventReviewService = new EventReviewService();
       $this->middleware('auth_volunteer', ['only' => [
           'create', 'store', 'edit', 'update', 'report'
       ]]);
@@ -24,12 +25,12 @@ class EventReviewAPIController extends Controller
 
   public function report($event_id, $review_id)
   {
-    $this->eventService->report($event_id,$review_id);
+      $this->eventReviewService->report($event_id,$review_id);
   }
 
   public function store(EventReviewRequest $request, $id)
   {
-    $this->eventService->store($request,$id);
+      $this->eventReviewService->store($request,$id);
   }
 
 }
