+++ app/Http/Controllers/API/OrganizationReviewAPIController.php
@@ -1,18 +1,20 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\API;
 
 use Illuminate\Http\Request;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests;
-use App\Http\Services\OrganizationService;
+use App\Http\Services\OrganizationReviewService;
 
 class OrganizationReviewAPIController extends Controller
 {
-  private $organizationService;
+  private $organizationReviewService;
+
   public function __construct()
   {
-      $this->organizationService = new OrganizationService();
+      $this->organizationReviewService = new OrganizationReviewService();
       $this->middleware('auth_volunteer', ['only' => [
           'create', 'store', 'edit', 'update', 'report'
       ]]);
@@ -20,12 +22,12 @@ class OrganizationReviewAPIController extends Controller
 
   public function store(OrganizationReviewRequest $request, $id)
   {
-      $this->organizationService->store($request,$id);
+      $this->organizationReviewService->store($request,$id);
   }
 
   public function report($organization_id, $review_id)
   {
-      $this->organizationService->report($organization_id,$review_id);
+      $this->organizationReviewService->report($organization_id,$review_id);
   }
 
 }
