+++ app/Http/Services/EventService.php
@@ -83,31 +83,4 @@ class EventService
     }
 
 
-    /*
-    |==========================================================================
-    | Event Review
-    |==========================================================================
-    |
-    */
-
-    /*
-    * Report event review.
-    */
-    public function report($event_id, $review_id)
-    {
-      $review = Event::findOrFail($event_id)->reviews()->findOrFail($review_id);
-      if(!$review->reportingUsers()->find(Auth::user()->id))
-          Auth::user()->reportedEventReviews()->attach($review);
-    }
-
-    /*
-    * Event Review.
-    */
-    public function store(EventReviewRequest $request, $id)
-    {
-          $review = new EventReview($request->all());
-          $review->user_id = Auth::user()->id;
-          $event = Event::findorfail($id);
-          $event->reviews()->save($review);
-    }
 }
